First brush at adding editorial comments to the calendar.
Click events aren't being dispatched properly in the item overlay -- related to selection not working. We'll have to fix that.

Comments themselves are just inline in the overlay for now (due to above). They should be moved to their own overlay that appears to the right of the main item overlay.

Barely any CSS yet to customize the output of comments. We will likely want to make editorial_comments_meta_box() more customizable with arguments so that we can avoid too much in the way of CSS hacks.

[#5930895]