Fix the status selection and the "Save as" link for Gutenberg
Don't show the select box for changing the status when the status is in the list: publish, future, pending, private. For all of them we show a static status name and a note with instructions to move to draft or disabling the pending checkbox. This makes it more consistent with Gutenberg UI because force users to use its UI to change from those statuses.

Removed Pending and Scheduled from the select field so users have to use the gutenberg fields for controlling that. Added support to display the Privately Published status, and fixed an issue in the “Save as …” link when the post was already published/scheduled. Now it shows correctly the “Save as Draft” to unpublish the post