Archive the notification log after it is sent
We archive instead of delete it because if the log is open on the screen and the user tries to preview its content, and on the same time the cron ran, if the cron delete the log the preview will fail and the user will think it is a bug. So we archive the log entry and don't show it on the log table.