If duplicate Notification Workflows were previously created, trash them.
If duplicates exist, they will have the standard post_name numerically suffixed with "-#"

It is also possible that after duplication, the original workflow was manually deleted. In that case, subsequent duplicates will use the original post_name, then the next available "-#" suffix.

Keep the oldest existing default workflow, even if it has a suffixed post_name. If this is not the copy the user wants deactivated, they can restore it from Trash.